IMN-357 - Moving protobuf converters to commons package (#282)
