--- packages/models/src/protobuf/index.ts
@@ -1,4 +1,0 @@
-export * from "./converterV1.js";
-export * from "./converterV2.js";
-export * from "./protobuf.js";
-export * from "./utils.js";
